+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/PrometheusExporterEndpointRouteBuilderExtensions.cs
@@ -37,7 +37,7 @@ namespace Microsoft.AspNetCore.Builder
         /// </summary>
         /// <remarks>Note: A branched pipeline is created for the route
         /// specified by <see
-        /// cref="PrometheusExporterOptions.ScrapeEndpointPath"/>.</remarks>
+        /// cref="PrometheusAspNetCoreOptions.ScrapeEndpointPath"/>.</remarks>
         /// <param name="endpoints">The <see cref="IEndpointRouteBuilder"/> to add
         /// middleware to.</param>
         /// <returns>A convention routes for the Prometheus scraping endpoint.</returns>
@@ -65,7 +65,7 @@ namespace Microsoft.AspNetCore.Builder
         /// <param name="endpoints">The <see cref="IEndpointRouteBuilder"/> to add
         /// middleware to.</param>
         /// <param name="path">Optional path to use for the branched pipeline.
-        /// If not provided then <see cref="PrometheusExporterOptions.ScrapeEndpointPath"/>
+        /// If not provided then <see cref="PrometheusAspNetCoreOptions.ScrapeEndpointPath"/>
         /// is used.</param>
         /// <param name="meterProvider">Optional <see cref="MeterProvider"/>
         /// containing a Prometheus exporter otherwise the primary SDK provider
@@ -83,16 +83,16 @@ namespace Microsoft.AspNetCore.Builder
             var builder = endpoints.CreateApplicationBuilder();
 
             // Note: Order is important here. MeterProvider is accessed before
-            // GetOptions<PrometheusExporterOptions> so that any changes made to
-            // PrometheusExporterOptions in deferred AddPrometheusExporter
+            // GetOptions<PrometheusAspNetCoreOptions> so that any changes made to
+            // PrometheusAspNetCoreOptions in deferred AddPrometheusExporter
             // configure actions are reflected.
             meterProvider ??= endpoints.ServiceProvider.GetRequiredService<MeterProvider>();
 
             if (path == null)
             {
-                var options = endpoints.ServiceProvider.GetRequiredService<IOptions<PrometheusExporterOptions>>().Value;
+                var options = endpoints.ServiceProvider.GetRequiredService<IOptions<PrometheusAspNetCoreOptions>>().Value;
 
-                path = options.ScrapeEndpointPath ?? PrometheusExporterOptions.DefaultScrapeEndpointPath;
+                path = options.ScrapeEndpointPath ?? PrometheusAspNetCoreOptions.DefaultScrapeEndpointPath;
             }
 
             if (!path.StartsWith("/"))